@@ -1,4 +1,6 @@
 <script src="https://unpkg.com/maplibre-gl@4.4.1/dist/maplibre-gl.js">
+	import { preventDefault } from 'svelte/legacy';
+
 	import { onMount } from 'svelte';
 	import maplibregl from 'maplibre-gl';
 	import Popup from './Popup.svelte';
@@ -12,18 +14,13 @@
 	let map;
 
 	/** @type {any} */
-	let popup_id;
+	let popup_id = $state();
 
-	/**
-	 * @type boolean
-	 */
-	export let showLeft = false;
-	/**
-	 * @type boolean
-	 */
-	export let showRight = false;
+	
+	
 
-	export let a = true;
+	/** @type {{showLeft?: boolean, showRight?: boolean, a?: boolean}} */
+	let { showLeft = $bindable(false), showRight = $bindable(false), a = true } = $props();
 
 	/** @type {string | number | undefined | null} */
 	let hoveredPointId = null;
@@ -32,13 +29,13 @@
 	let hoveredLineId = null;
 
 	// Calculate the width and left values based on the presence of the left and right pop-ups
-	$: mapWidth =
-		showLeft && showRight
+	let mapWidth =
+		$derived(showLeft && showRight
 			? 'calc(100% - 52.08vw)'
 			: showLeft || showRight
 				? 'calc(100% - 26.04vw)'
-				: '100%';
-	$: mapLeft = showLeft ? '26.04vw' : '0';
+				: '100%');
+	let mapLeft = $derived(showLeft ? '26.04vw' : '0');
 
 	// Functions to handle the custom events
 	/**
@@ -377,7 +374,7 @@
 				><input
 					type="checkbox"
 					bind:checked={$isMapOChecked}
-					on:click|preventDefault={(event) => handleClick(event, 'MapO')}
+					onclick={preventDefault((event) => handleClick(event, 'MapO'))}
 				/>
 				seuls emplacements <span style="color: red;">actuels</span> / nur
 				<span style="color: red;">aktuelle</span> Standorte</label
@@ -388,13 +385,13 @@
 				><input
 					type="checkbox"
 					bind:checked={$isMapNChecked}
-					on:click|preventDefault={(event) => handleClick(event, 'MapN')}
+					onclick={preventDefault((event) => handleClick(event, 'MapN'))}
 				/>
 				seuls emplacements <span style="color: blue;">d'origine</span> / nur
 				<span style="color: blue;">ursprüngliche</span> Standorte</label
 			>
 			<br />
-			<button on:click={resetMap} style="width: 100%;"
+			<button onclick={resetMap} style="width: 100%;"
 				>réinitialiser la carte / Karte zurücksetzen</button
 			>
 		</div>
